requirements.txt
attrs==22.2.0
cached-property==1.3.1
gdown==5.1.0
ipython==7.34.0
ipywidgets==8.0.2
jupyter==1.0.0
matplotlib==3.1.1
numpy==1.23.5
opencv-contrib-python==4.5.5.64
pythreejs==2.4.2
torch==2.0.1
tqdm==4.64.0
